+++ rust/c509-certificate/src/algorithm_identifier.rs
@@ -17,7 +17,7 @@ use asn1_rs::Oid;
 use minicbor::{encode::Write, Decode, Decoder, Encode, Encoder};
 use serde::{Deserialize, Serialize};
 
-use crate::c509_oid::C509oid;
+use crate::oid::C509oid;
 
 /// A struct represents the `AlgorithmIdentifier` type.
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
