+++ rust/c509-certificate/src/lib.rs
@@ -40,20 +40,20 @@ use c509::C509;
 use minicbor::{Decode, Encode};
 use signing::{PrivateKey, PublicKey};
 use tbs_cert::TbsCert;
+pub mod algorithm_identifier;
+pub mod attributes;
+pub mod big_uint;
 pub mod c509;
-pub mod c509_algo_identifier;
-pub mod c509_attributes;
-pub mod c509_big_uint;
-pub mod c509_extensions;
-pub mod c509_general_names;
-pub mod c509_issuer_sig_algo;
-pub mod c509_name;
-pub mod c509_oid;
-pub mod c509_subject_pub_key_algo;
-pub mod c509_time;
+pub mod extensions;
+pub mod general_names;
+pub mod issuer_sig_algo;
+pub mod name;
+pub mod oid;
 pub mod signing;
+pub mod subject_pub_key_algo;
 mod tables;
 pub mod tbs_cert;
+pub mod time;
 pub mod wasm_binding;
 
 /// Generate a signed or unsigned C509 certificate.
