+++ rust/c509-certificate/src/subject_pub_key_algo/data.rs
